use feature 'state' instead of 5.016
A lot of people still don't have Perl 5.16. 5.10 is pretty safe, 
though. You are using `use 5.016` simply to enable the `state` 
keyword, but that was available as of 5.10, so you could `use 5.010`. 
However, rather than enabling all of the features of Perl 5.10, it's 
much safer to just enable what you need, which would mean simply
`use feature 'state';`.